ib-api/requirements.txt
# FastAPI and ASGI server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Interactive Brokers TWS API
ibapi

# JWT Authentication
python-jose[cryptography]>=3.3.0

# Password hashing
passlib[bcrypt]>=1.7.4

# Settings management
pydantic-settings>=2.0.0
